Move brew specific functions to specific/brew.py
